removes repositories drop down
